+++ backend/src/main/java/com/dopamines/backend/plan/repository/PlanRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 
-public interface PlanRepository extends JpaRepository<Plan, Long> {
+public interface PlanRepository extends JpaRepository<Plan, Integer> {
     Optional<Plan> findById(Integer planId);
 }
